refactor(web): extract useContextWindowUsageHydration (LUM-1740 Phase 5a)
Phase 5 of the chat-page data lifecycle redesign starts dismantling
use-conversation-loader.ts and folding its residuals into focused hooks.
This step extracts the ~20-line hydration effect — read the per-assistant
map from localStorage on assistant change and seed the active
conversation's indicator — into its own hook mounted directly by ChatPage.

Net: loader drops the hydratedAssistantIdRef bookkeeping and
loadContextWindowUsageMap import. The page-level
contextWindowUsageByConversationRef still lives in chat-page (until the
remaining ref-sync block goes away in the rest of Phase 5).
